small fixes

+++ starknet-crypto/src/ecdsa.rs
@@ -206,7 +206,7 @@ pub fn recover(message: &Felt, r: &Felt, s: &Felt, v: &Felt) -> Result<Felt, Rec
 #[inline(always)]
 fn mul_by_bits(x: &AffinePoint, y: &Felt) -> AffinePoint {
     let x = ProjectivePoint::from(x);
-    let mut y_bool = [false; 256 as usize];
+    let mut y_bool = [false; 256_usize];
     for (bool_ref, bit) in y_bool.iter_mut().zip(y.to_bits_le().iter().by_vals()) {
         *bool_ref = bit;
     }
